# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: you@your-company.com
  description: This is a simple API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AL
  version: 1.0.1
servers:
- description: SwaggerHub API Auto Mocking
  url: https://accelerated-life-demo.herokuapp.com/api
...
paths:
...
  /schools/school:
    delete:
      description: delete a school
      operationId: deleteSchool
      parameters:
      - description: pass the _id to be deleted
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          description: School Deleted
        "400":
          description: School cannot be edited with that data
      summary: delete school
      tags:
      - SysAdmin
    get:
      description: |
        By passing in a _id you can find a school
      operationId: searchSchool
      parameters:
      - description: pass the _id to be searched
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/School'
                type: array
          description: school matching criteria
        "404":
          description: No school found
      summary: searches for a school
      tags:
      - All
    post:
      description: creates a new school
      operationId: makeSchool
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/schools_school_body_1'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
          description: school record is made
        "400":
          description: School cannot be made with that data
      summary: make a new school
      tags:
      - SysAdmin
    put:
      description: edit a school
      operationId: editSchool
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/schools_school_body'
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/School'
                type: array
          description: school record is made
        "400":
          description: School cannot be edited with that data
      summary: edit school
      tags:
      - SysAdmin
  /accounts/account:
    delete:
      description: delete an account
      operationId: deleteAccount
      parameters:
      - description: pass the _id to be deleted
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          description: Account Deleted
        "400":
          description: Account cannot be edited with that data
      summary: delete school
      tags:
      - SysAdmin
    get:
      description: By passing in a _id you can find an Account
      operationId: searchAccount
      parameters:
      - description: pass the _id to be searched
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Account'
                type: array
          description: account matching criteria
        "404":
          description: No account found
      summary: searches for a account
      tags:
      - All
    post:
      description: make an account
      operationId: makeAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/accounts_account_body_1'
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Account'
                type: array
          description: school record is made
        "400":
          description: Account cannot be edited with that data
      summary: make account
      tags:
      - SysAdmin
    put:
      description: edit an account
      operationId: editAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/accounts_account_body'
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Account'
                type: array
          description: school record is made
        "400":
          description: Account cannot be edited with that data
      summary: edit account
      tags:
      - SysAdmin
...
  /accounts/all:
    get:
      description: users user._id to find all buck accounts for that user. Need to
        also find its conversion with the school UBuck
      operationId: searchStudentBucks
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ResponseCurrencyExchange'
                type: array
          description: accounts matching criteria
        "404":
          description: No accounts found
      summary: returns all buck accounts for specific user
      tags:
      - All
  /accounts/exchangeRate:
    get:
      description: given sellCurrency and buyCurrency return the exchange rate between
        the two currencies. Also return the students details on each account.
      operationId: exchangeRate
      parameters:
      - description: the _id of the currency to be sold
        explode: true
        in: query
        name: sellCurrency
        required: true
        schema:
          type: string
        style: form
      - description: the _id of the currency to be bought
        explode: true
        in: query
        name: buyCurrency
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ResponseCurrencyExchange'
                type: array
          description: accounts with names and conversion details
        "404":
          description: No accounts found
      summary: returns exchange rate between 2 buck accounts
      tags:
      - All
...
  /bucks/buck:
    delete:
      description: delete a buck. All accounts with that buck should first have value
        converted to UBucks and then deleted.
      operationId: deleteBuck
      parameters:
      - description: pass the _id to be deleted
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Buck'
          description: Buck record is edited
        "404":
          description: Buck cannot be edited with that data
      summary: delete buck
      tags:
      - SysAdmin
    get:
      description: by passing in a buck ID that buck is returned
      operationId: searchBucks
      parameters:
      - description: pass the id to be searched
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Buck'
          description: buck matching criteria
        "404":
          description: No buck matching that _id
      summary: searches bucks
      tags:
      - All
    post:
      description: create a buck
      operationId: createBuck
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bucks_buck_body_1'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Buck'
          description: Buck record is created
        "404":
          description: Buck cannot be created with that data
      summary: create buck
      tags:
      - SysAdmin
    put:
      description: edit a buck
      operationId: editBuck
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bucks_buck_body'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Buck'
          description: Buck record is edited
        "404":
          description: Buck cannot be edited with that data
      summary: edit buck
      tags:
      - SysAdmin
  /bucks:
    get:
      description: return all the bucks of the school
      operationId: searchAllBucks
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Buck'
                type: array
          description: bucks matching criteria in alphabetical order decending
        "404":
          description: No bucks matching that school _id
      summary: searches all bucks
      tags:
      - All
  /classes/class:
    delete:
      description: delete a class
      operationId: deleteclass
      parameters:
      - description: pass the _id to be deleted
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          description: School Deleted
        "404":
          description: Class cannot be deleted with that data
      summary: delete class
      tags:
      - Staff
    get:
      description: By passing in a _id you can find a class. Members should be populated
        with firstName, lastName, rank, netWorth
      operationId: searchClass
      parameters:
      - description: pass the _id to be searched
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassWithMembers'
          description: class matching criteria
        "404":
          description: No class found
      summary: searches for a class
      tags:
      - All
    post:
      description: creates a new class
      operationId: makeClass
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestMakeClass'
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Class'
                type: array
          description: class record is made. Return all the classes from this owner.
            Sort classes by period ascending
        "404":
          description: Class cannot be made with that data
      summary: make a new class
      tags:
      - Staff
    put:
      description: edit a class. return all the classes of the owner. Exclude owner,
        createdAt, updatedAt and version. Sort by period ascending
      operationId: editClass
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestEditClass'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
          description: school record is edited
        "404":
          description: Class cannot be edited with that data
      summary: edit class
      tags:
      - Staff
...
  /users:
    get:
      description: sort by rank ascending
      operationId: searchStudents
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserNoHistory'
                type: array
          description: students matching criteria
        "404":
          description: No users found with that data
      summary: return all students from a school
      tags:
      - All
  /users/user:
    delete:
      description: Delete a student.
      operationId: deleteStudent
      parameters:
      - description: pass the email to be deleted. Delete all of their transactions
          and events. Search classes that the user is a member in and remove them
          from the member list. Find all of their buck accounts and burn the bucks
          then delete the account. Find all the crypto accounts, modify the cryptos
          totalCount then delete the account
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          description: transaction Deleted
        "404":
          description: There is no user with that email
      summary: delete student
      tags:
      - Staff
    get:
      description: return student
      operationId: searchStudent
      parameters:
      - description: pass the _id to be searched
        explode: true
        in: query
        name: _id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: students matching criteria
        "404":
          description: No user found with that data
      summary: return one student
      tags:
      - All
    put:
      description: edit user fields. If it is password then the password must be salted
        and hashed. If it is college or career then income needs to be halved and
        end times need to be saved to the record. Email change means verified is set
        to false. Request will have new password in the clear. If college true then
        transaction has to be made representing the cost of college.
      operationId: userEdit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/users_user_body'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: return user information with history
        "404":
          description: Cannot update user with that data
      summary: edit a user
      tags:
      - All
  /users/auth:
    get:
      description: return user
      operationId: AuthUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseAuth2'
          description: students matching criteria
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseAuth4'
          description: No user found with that data
      summary: return authenticated user details
      tags:
      - All
...
  /users/login:
    post:
      description: If user email and password are correc tthen generate token. If
        use is not confirmed then send them an email with confirmation URL
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestLogin'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseLogin2'
          description: class record is made. Return all the classes from this owner.
            Sort classes by period ascending
          headers:
            Set-Cookie:
              description: Contains the session cookie named `w_authExp` and `w_auth`.
                Pass this cookie back in subsequent requests.
              explode: false
              schema:
                $ref: '#/components/schemas/ResponseCookie'
              style: simple
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseLogin4'
          description: wrong password
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseLogin4'
          description: wrong password
      summary: logging in
      tags:
      - All
  /users/logout:
    get:
      description: logout
      operationId: logout
      parameters:
      - description: pass the _id to be searched
        explode: false
        in: header
        name: _id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseRegister2'
          description: students matching criteria
        "404":
          description: No user found with that data
      summary: logout
      tags:
      - All
...
  /users/confirmEmail:
    get:
      description: user is authenticated by token and confirmed is changed to true
      operationId: confirmEmail
      parameters:
      - description: pass the token to be searched
        explode: true
        in: query
        name: token
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_4'
          description: students matching criteria
        "404":
          description: No user found with that data
      summary: confirm email
      tags:
      - All
components:
  schemas:
    Account:
      example:
        owner_id: 613eab19e58cc511801af657
        kind: Buck
        type_id: 613eab19e58cc511801af657
        _id: 613eab19e58cc511801af657
        history:
        - - 0
          - 0
        - - 0
          - 0
        basis: 1245.7
        value: 1245.7
      properties:
        _id:
          example: 613eab19e58cc511801af657
          type: string
        owner_id:
          example: 613eab19e58cc511801af657
          type: string
        kind:
          example: Buck
          type: string
        type_id:
          example: 613eab19e58cc511801af657
          type: string
        value:
          default: 0
          example: 1245.7
          format: float
          type: number
        basis:
          example: 1245.7
          format: float
          type: number
        history:
          items:
            items:
              type: integer
            maxItems: 2
            minItems: 2
            type: array
          type: array
      required:
      - _id
      - basis
      - history
      - kind
      - owner
      - type
      - value
      type: object
...
    Buck:
      example:
        name: UBuck
        _id: 613eab19e58cc511801af657
        ratio: 0.8008282
      properties:
        _id:
          example: 613eab19e58cc511801af657
          type: string
        name:
          example: UBuck
          type: string
        ratio:
          format: float
          type: number
      required:
      - _id
      - name
      - ratio
      type: object
...
    ClassWithMembers:
      example:
        period: 4
        owner_id: 613eab19e58cc511801af657
        members:
        - firstName: Jay
          lastName: Leno
          netWorth: 22.1
          rank: 4
          _id: 613eab19e58cc511801af657
        - firstName: Jay
          lastName: Leno
          netWorth: 22.1
          rank: 4
          _id: 613eab19e58cc511801af657
        name: Math
        _id: 613eab19e58cc511801af657
        addCode: Sd45Tb
      properties:
        _id:
          example: 613eab19e58cc511801af657
          type: string
        owner_id:
          example: 613eab19e58cc511801af657
          type: string
        period:
          example: 4
          type: integer
        name:
          example: Math
          type: string
        addCode:
          example: Sd45Tb
          type: string
        members:
          items:
            $ref: '#/components/schemas/ClassWithMembers_members'
          type: array
      required:
      - _id
      - addCode
      - members
      - name
      - owner_id
      - period
      type: object
...
    School:
      example:
        zip: 95338
        city: Stockton
        adminAddCode: adc24F
        name: Edison High School
        _id: 613eab19e58cc511801af657
      properties:
        _id:
          example: 613eab19e58cc511801af657
          type: string
        name:
          example: Edison High School
          type: string
        city:
          example: Stockton
          type: string
        adminAddCode:
          example: adc24F
          type: string
        zip:
          example: 95338
          type: integer
      required:
      - _id
      - adminAddCode
      - city
      - name
      - zip
      type: object
...
    ResponseAuth4:
      properties:
        isAuth:
          example: false
          type: boolean
        error:
          type: boolean
      type: object
    ResponseCookie:
      properties:
        w_authExp:
          type: string
        w_auth:
          type: string
      type: object
...
    ResponseCurrencyExchange:
      example:
        balance: 123.45
        buck:
          name: Jones Bucks
          history:
          - date: 2000-01-23T04:56:07.000+00:00
            netWorth: 123.45
          - date: 2000-01-23T04:56:07.000+00:00
            netWorth: 123.45
        type_id: 614290257145de7a9e40c594
        history:
        - date: 2000-01-23T04:56:07.000+00:00
          netWorth: 123.45
        - date: 2000-01-23T04:56:07.000+00:00
          netWorth: 123.45
        _id: 614290257145de7a9e40c594
        conversion: 123.45
      properties:
        conversion:
          example: 123.45
          format: float
          type: number
        history:
          items:
            $ref: '#/components/schemas/History'
          type: array
        buck:
          $ref: '#/components/schemas/ResponseCurrencyExchange_buck'
        balance:
          example: 123.45
          format: float
          type: number
        _id:
          example: 614290257145de7a9e40c594
          type: string
        type_id:
          example: 614290257145de7a9e40c594
          type: string
      required:
      - balance
      - conversion
      type: object
...
    RequestEditClass:
      example:
        period: 4
        name: Math
        _id: 613eab19e58cc511801af657
      properties:
        name:
          example: Math
          type: string
        period:
          example: 4
          type: integer
        _id:
          example: 613eab19e58cc511801af657
          type: string
      type: object
...
    RequestMakeClass:
      example:
        period: 4
        owner_id: 613eab19e58cc511801af657
        name: Math
      properties:
        name:
          example: Math
          type: string
        owner_id:
          example: 613eab19e58cc511801af657
          type: string
        period:
          example: 4
          type: integer
      type: object
...
    schools_school_body:
      properties:
        name:
          example: East High
          type: string
        city:
          example: Dublin
          type: string
        zip:
          example: 94557
          type: integer
      type: object
    schools_school_body_1:
      properties:
        name:
          example: East High
          type: string
        city:
          example: Dublin
          type: string
        zip:
          example: 94557
          type: integer
      type: object
    inline_response_200:
      example:
        success: true
      properties:
        success:
          description: True if record was created
          type: boolean
      type: object
    accounts_account_body:
      properties:
        _id:
          example: 614290257145de7a9e40c594
          type: string
        owner:
          example: 614290257145de7a9e40c594
          type: string
        kind:
          example: Buck
          type: string
        type:
          example: 613eab19e58cc511801af657
          type: string
        value:
          example: 123.45
          format: float
          type: number
        basis:
          example: 123.45
          format: float
          type: number
      type: object
    accounts_account_body_1:
      properties:
        owner:
          example: 614290257145de7a9e40c594
          type: string
        kind:
          example: Buck
          type: string
        type:
          example: 613eab19e58cc511801af657
          type: string
        value:
          example: 123.45
          format: float
          type: number
        basis:
          example: 123.45
          format: float
          type: number
      type: object
...
    bucks_buck_body:
      properties:
        name:
          example: Jane Bucks
          type: string
        owner:
          example: 613eab19e58cc511801af657
          type: string
        school:
          example: 613eab19e58cc511801af657
          type: string
        totalCurrency:
          example: 123.45
          format: float
          type: number
        freeCurrency:
          example: 123.45
          format: float
          type: number
      type: object
    bucks_buck_body_1:
      properties:
        name:
          example: UBuck
          type: string
        owner:
          example: 613eab19e58cc511801af657
          type: string
        school:
          example: 613eab19e58cc511801af657
          type: string
      type: object
...